refactor: Product, headquarter 패키지에 리팩토링
