+++ src/main/java/uos/uos25/headquarter/dto/request/ProductInfoRequestDTO.java
@@ -1,4 +1,4 @@
-package uos.uos25.product.dto.request;
+package uos.uos25.headquarter.dto.request;
 
 import lombok.Data;
