+++ src/main/java/uos/uos25/headquarter/dto/response/ProductCreateResponseDTO.java
@@ -1,10 +1,10 @@
-package uos.uos25.product.dto.response;
+package uos.uos25.headquarter.dto.response;
 
 import java.time.LocalDateTime;
 
 import lombok.Builder;
 import lombok.Data;
-import uos.uos25.product.entity.Product;
+import uos.uos25.headquarter.entity.Product;
 
 @Data
 public class ProductCreateResponseDTO {
