+++ src/main/java/uos/uos25/headquarter/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package uos.uos25.product.repository;
+package uos.uos25.headquarter.repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import uos.uos25.product.entity.Product;
+import uos.uos25.headquarter.entity.Product;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, String> {
